requirements.txt
amqp==2.1.4
beautifulsoup4==4.5.1
billiard==3.5.0.2
blinker==1.4
celery==4.0.2
click==6.7
Flask==0.12.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
-e git+https://github.com/radusuciu/ip2api.git@3626fa5e2bab79154449ea8ec0a2e23f0232a782#egg=ip2api
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.0.2
ldap3==2.2.4
MarkupSafe==1.0
marshmallow==2.13.5
passlib==1.7.1
pkg-resources==0.0.0
pyasn1==0.2.3
pytz==2017.2
PyYAML==3.12
requests==2.14.2
SQLAlchemy==1.1.9
vine==1.1.3
Werkzeug==0.12.2
WTForms==2.1
